requirements.txt
aiohttp==3.13.3
altair==6.0.0
anthropic==0.84.0
chromadb==1.5.2
fastapi==0.135.1
folium==0.20.0
groq==0.37.1
httpx==0.28.1
langchain==1.2.10
langchain-anthropic==1.3.4
langchain-community==0.4.1
langchain-core==1.2.17
langchain-groq==1.1.2
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
langgraph==1.0.10
langgraph-prebuilt==1.0.8
mcp==1.26.0
numpy==2.4.2
openai==2.26.0
pandas==2.3.3
pydantic==2.12.5
pydantic-settings==2.13.1
pypdf==6.7.5
python-dotenv==1.2.2
sentence-transformers==5.2.3
streamlit==1.55.0
streamlit-folium==0.26.2
uvicorn==0.41.0
